Simplify TypeScript types package
