+++ .vscode/settings.json
@@ -4,5 +4,6 @@
         "source.organizeImports",
         "source.fixAll"
     ],
-    "typescript.tsdk": "node_modules/typescript/lib"
+    "typescript.tsdk": "node_modules/typescript/lib",
+    "typescript.preferences.importModuleSpecifier": "relative"
 }
