+++ package.json
@@ -32,8 +32,9 @@
         "shx": "^0.3.2",
         "sinon": "^6.0.0",
         "string-replace-webpack-plugin": "^0.1.3",
-        "ts-node": "^3.3.0",
         "typescript": "^4.1.3",
+        "typescript3": "npm:typescript@~3.0.0",
+        "typescript4": "npm:typescript@~4.0.0",
         "webpack": "^3.12.0"
     },
     "homepage": "https://github.com/Azure/azure-functions-nodejs-worker",
@@ -46,13 +47,13 @@
     },
     "scripts": {
         "clean": "rimraf dist && rimraf azure-functions-language-worker-protobuf/src/rpc*",
-        "build": "rimraf dist && npm run gen && shx mkdir -p dist/azure-functions-language-worker-protobuf/src && shx cp azure-functions-language-worker-protobuf/src/rpc.* dist/azure-functions-language-worker-protobuf/src/. && tsc",
+        "build": "rimraf dist && npm run gen && shx mkdir -p dist/azure-functions-language-worker-protobuf/src && shx cp azure-functions-language-worker-protobuf/src/rpc.* dist/azure-functions-language-worker-protobuf/src/. && node ./node_modules/typescript/bin/tsc",
         "build-nomaps": "npm run build -- --sourceMap false",
         "gen": "node scripts/generateProtos.js",
-        "test": "mocha -r ts-node/register ./test/**/*.ts",
+        "test": "mocha ./dist/test/**/*.js",
         "lint": "eslint src",
         "lint-fix": "eslint src --fix",
-        "watch": "tsc --watch",
+        "watch": "node ./node_modules/typescript/bin/tsc --watch",
         "prepare": "husky install"
     },
     "files": [
