+++ src/WorkerChannel.ts
@@ -1,3 +1,4 @@
+import { Context } from '@azure/functions';
 import { format, isFunction } from 'util';
 import { AzureFunctionsRpcMessages as rpc } from '../azure-functions-language-worker-protobuf/src/rpc';
 import { augmentTriggerMetadata } from './augmenters';
@@ -5,7 +6,6 @@ import { CreateContextAndInputs, LogCallback, ResultCallback } from './Context';
 import { toTypedData } from './converters';
 import { IFunctionLoader } from './FunctionLoader';
 import { IEventStream } from './GrpcService';
-import { Context } from './public/Interfaces';
 import { InternalException } from './utils/InternalException';
 import { systemError, systemWarn } from './utils/Logger';
 import LogCategory = rpc.RpcLog.RpcLogCategory;
