+++ src/http/Request.ts
@@ -1,6 +1,6 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
-import { HttpMethod, HttpRequest } from '../public/Interfaces';
+import { HttpMethod, HttpRequest } from '@azure/functions';
 
 export class RequestProperties implements HttpRequest {
     method: HttpMethod | null = null;
