--- src/public/Interfaces.ts
@@ -1,300 +1,0 @@
-/**
- * Context bindings object. Provided to your function binding data, as defined in function.json.
- */
-export interface ContextBindings {
-    [name: string]: any;
-}
-
-/**
- * Context binding data. Provided to your function trigger metadata and function invocation data.
- */
-export interface ContextBindingData {
-    invocationId?: string | null;
-    [name: string]: any;
-}
-
-/**
- * The context object can be used for writing logs, reading data from bindings, setting outputs and using
- * the context.done callback when your exported function is synchronous. A context object is passed
- * to your function from the Azure Functions runtime on function invocation.
- */
-export interface Context {
-    /**
-     * A unique GUID per function invocation.
-     */
-    invocationId: string;
-    /**
-     * Function execution metadata.
-     */
-    executionContext: ExecutionContext;
-    /**
-     * Input and trigger binding data, as defined in function.json. Properties on this object are dynamically
-     * generated and named based off of the "name" property in function.json.
-     */
-    bindings: ContextBindings;
-    /**
-     * Trigger metadata and function invocation data.
-     */
-    bindingData: ContextBindingData;
-    /**
-     * TraceContext information to enable distributed tracing scenarios.
-     */
-    traceContext: TraceContext;
-    /**
-     * Bindings your function uses, as defined in function.json.
-     */
-    bindingDefinitions: BindingDefinition[];
-    /**
-     * Allows you to write streaming function logs. Calling directly allows you to write streaming function logs
-     * at the default trace level.
-     */
-    log: Logger;
-    /**
-     * A callback function that signals to the runtime that your code has completed. If your function is synchronous,
-     * you must call context.done at the end of execution. If your function is asynchronous, you should not use this
-     * callback.
-     *
-     * @param err A user-defined error to pass back to the runtime. If present, your function execution will fail.
-     * @param result An object containing output binding data. `result` will be passed to JSON.stringify unless it is
-     *  a string, Buffer, ArrayBufferView, or number.
-     */
-    done(err?: Error | string | null, result?: any): void;
-    /**
-     * HTTP request object. Provided to your function when using HTTP Bindings.
-     */
-    req?: HttpRequest;
-    /**
-     * HTTP response object. Provided to your function when using HTTP Bindings.
-     */
-    res?: { [key: string]: any };
-}
-
-/**
- * HTTP request headers.
- */
-export interface HttpRequestHeaders {
-    [name: string]: string | undefined;
-}
-
-/**
- * Query string parameter keys and values from the URL.
- */
-export interface HttpRequestQuery {
-    [name: string]: string | undefined;
-}
-
-/**
- * Route parameter keys and values.
- */
-export interface HttpRequestParams {
-    [name: string]: string | undefined;
-}
-
-/**
- * HTTP request object. Provided to your function when using HTTP Bindings.
- */
-export interface HttpRequest {
-    /**
-     * HTTP request method used to invoke this function.
-     */
-    method: HttpMethod | null;
-    /**
-     * Request URL.
-     */
-    url: string;
-    /**
-     * HTTP request headers.
-     */
-    headers: HttpRequestHeaders;
-    /**
-     * Query string parameter keys and values from the URL.
-     */
-    query: HttpRequestQuery;
-    /**
-     * Route parameter keys and values.
-     */
-    params: HttpRequestParams;
-    /**
-     * The HTTP request body.
-     */
-    body?: any;
-    /**
-     * The HTTP request body as a UTF-8 string.
-     */
-    rawBody?: any;
-}
-
-/**
- * Possible values for an HTTP request method.
- */
-export type HttpMethod = 'GET' | 'POST' | 'DELETE' | 'HEAD' | 'PATCH' | 'PUT' | 'OPTIONS' | 'TRACE' | 'CONNECT';
-
-/**
- * Http response cookie object to "Set-Cookie"
- */
-export interface Cookie {
-    /** Cookie name */
-    name: string;
-
-    /** Cookie value */
-    value: string;
-
-    /** Specifies allowed hosts to receive the cookie */
-    domain?: string;
-
-    /** Specifies URL path that must exist in the requested URL */
-    path?: string;
-
-    /**
-     * NOTE: It is generally recommended that you use maxAge over expires.
-     * Sets the cookie to expire at a specific date instead of when the client closes.
-     * This can be a Javascript Date or Unix time in milliseconds.
-     */
-    expires?: Date | number;
-
-    /** Sets the cookie to only be sent with an encrypted request */
-    secure?: boolean;
-
-    /** Sets the cookie to be inaccessible to JavaScript's Document.cookie API */
-    httpOnly?: boolean;
-
-    /** Can restrict the cookie to not be sent with cross-site requests */
-    sameSite?: 'Strict' | 'Lax' | 'None' | undefined;
-
-    /** Number of seconds until the cookie expires. A zero or negative number will expire the cookie immediately. */
-    maxAge?: number;
-}
-
-export interface ExecutionContext {
-    /**
-     * A unique GUID per function invocation.
-     */
-    invocationId: string;
-    /**
-     * The name of the function that is being invoked. The name of your function is always the same as the
-     * name of the corresponding function.json's parent directory.
-     */
-    functionName: string;
-    /**
-     * The directory your function is in (this is the parent directory of this function's function.json).
-     */
-    functionDirectory: string;
-    /**
-     * The retry context of the current funciton execution. The retry context of the current function execution. Equals null if retry policy is not defined or it's the first function execution.
-     */
-    retryContext?: RetryContext;
-}
-
-export interface RetryContext {
-    /**
-     * Current retry count of the function executions.
-     */
-    retryCount: number;
-    /**
-     * Max retry count is the maximum number of times an execution is retried before eventual failure. A value of -1 means to retry indefinitely.
-     */
-    maxRetryCount: number;
-    /**
-     * Exception that caused the retry
-     */
-    exception?: Exception;
-}
-
-export interface Exception {
-    /** Exception source */
-    source?: string | null;
-
-    /** Exception stackTrace */
-    stackTrace?: string | null;
-
-    /** Exception message */
-    message?: string | null;
-}
-
-/**
- * TraceContext information to enable distributed tracing scenarios.
- */
-export interface TraceContext {
-    /** Describes the position of the incoming request in its trace graph in a portable, fixed-length format. */
-    traceparent: string | null | undefined;
-
-    /** Extends traceparent with vendor-specific data. */
-    tracestate: string | null | undefined;
-
-    /** Holds additional properties being sent as part of request telemetry. */
-    attributes:
-        | {
-              [k: string]: string;
-          }
-        | null
-        | undefined;
-}
-
-export interface BindingDefinition {
-    /**
-     * The name of your binding, as defined in function.json.
-     */
-    name: string;
-    /**
-     * The type of your binding, as defined in function.json.
-     */
-    type: string;
-    /**
-     * The direction of your binding, as defined in function.json.
-     */
-    direction: 'in' | 'out' | 'inout' | undefined;
-}
-
-/**
- * Allows you to write streaming function logs.
- */
-export interface Logger {
-    /**
-     * Writes streaming function logs at the default trace level.
-     */
-    (...args: any[]): void;
-    /**
-     * Writes to error level logging or lower.
-     */
-    error(...args: any[]): void;
-    /**
-     * Writes to warning level logging or lower.
-     */
-    warn(...args: any[]): void;
-    /**
-     * Writes to info level logging or lower.
-     */
-    info(...args: any[]): void;
-    /**
-     * Writes to verbose level logging.
-     */
-    verbose(...args: any[]): void;
-}
-
-export interface Timer {
-    isPastDue: boolean;
-    schedule: {
-        /**
-         * Describes whether the timer for daylight saving based on the timer's time zone
-         */
-        adjustForDST: boolean;
-        /**
-         * Describes a timer schedule status
-         */
-        /** Describes whether the timer schedules is past it's scheduled due time */
-    };
-    scheduleStatus: {
-        /**
-         * Describes the last recorded schedule occurence. Date ISO string.
-         */
-        last: string;
-        /**
-         * Describes the expected next schedule occurrence
-         */
-        next: string;
-        /**
-         * Describes the last time this record was updated. This is used to re-calculate `next` with the current schedule after a host restart. Date ISO string.
-         */
-        lastUpdated: string;
-    };
-}
