--- types/public/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@azure/functions",
-  "version": "2.1.2",
-  "description": "Azure Functions types for Typescript",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Azure/azure-functions-nodejs-worker/tree/v2.x/types/public"
-  },
-  "keywords": [
-    "azure-functions",
-    "typescript"
-  ],
-  "types": "./Main.d.ts",
-  "typesVersions": {
-    ">=3.1": {
-      "*": [
-        "ts3.1/*"
-      ]
-    }
-  },
-  "author": "Microsoft",
-  "license": "MIT"
-}
